requirements.txt
requests>=2.31.0
typing-extensions>=4.7.0
flask>=2.3.0
flask-cors>=4.0.0
groq
pillow
python-dotenv 